image-video-generation-lab/requirements.txt
# Core UI / ML
gradio>=4.44.0
diffusers>=0.29.0
transformers>=4.42.0
accelerate>=0.33.0
safetensors>=0.4.4

# Media
pillow>=10.3.0
numpy>=1.26.4
moviepy>=1.0.3
pyttsx3>=2.90

# Utils
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.66.4
srt>=3.5.3
